requirements.txt
# Server framework
flask

# Sound processing
numpy
pysoundfile
librosa
webrtcvad

# Authentication
jwcrypto
pyjwt

# Data storage
PyMySQL
pandas

# PDF Generation
jinja2
weasyprint

# AWS Interaction
botocore
boto3

# Tests
pytest